refactor store subpackage to export from base, rather than nested 'store'
