+++ src/mlte/artifact/artifact.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 from mlte.artifact.model import ArtifactModel, ArtifactType
 from mlte.context import Context
 from mlte.session import session
-from mlte.store import Store
+from mlte.store.base import Store
 
 
 class Artifact:
