+++ test/store/fixture.py
@@ -13,7 +13,7 @@ from fastapi.testclient import TestClient
 
 import mlte.web.store.app_factory as app_factory
 from mlte.artifact.model import ArtifactType
-from mlte.store import StoreURI
+from mlte.store.base import StoreURI
 from mlte.store.factory import create_store
 from mlte.store.underlying.fs import LocalFileSystemStore
 from mlte.store.underlying.http import (
